Flatten pbnh.db
